Backend/pyproject.toml
[project]
name = "youtube-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.33.0",
    "openai-agents>=0.4.2",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
]
